pyproject.toml
[project]
name = "elon"
version = "0.0.1"
description = "Autonomous brand marketing agent"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # DB
    "sqlalchemy>=2.0.30",
    "alembic>=1.13",
    "psycopg[binary,pool]>=3.2",
    "pgvector>=0.3",
    # Queue
    "celery[redis]>=5.4",
    "redis>=5.0",
    # Storage
    "boto3>=1.35",
    # LLM
    "anthropic>=0.40",
    # Chat
    "python-telegram-bot>=21.4",
    "httpx>=0.27",
    # Crypto
    "pynacl>=1.5",
    # Scraping / media
    "playwright>=1.46",
    "beautifulsoup4>=4.12",
    "pillow>=10.4",
    "colorthief>=0.2.1",
    "pdfplumber>=0.11",
    # Observability
    "structlog>=24.4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "sentry-sdk[fastapi]>=2.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires DATABASE_URL pointing at Postgres+pgvector",
]
